backend/requirements.txt
annotated-types    0.5.0
anyio              3.7.1
asttokens          2.2.1
attrs              23.1.0
backcall           0.2.0
certifi            2023.7.22
cffi               1.15.1
charset-normalizer 3.2.0
click              8.1.6
comm               0.1.4
cryptography       41.0.3
debugpy            1.6.7.post1
decorator          5.1.1
ecdsa              0.18.0
exceptiongroup     1.1.2
executing          1.2.0
fastapi            0.101.0
greenlet           2.0.2
h11                0.14.0
httpcore           0.17.3
httpx              0.24.1
idna               3.4
iniconfig          2.0.0
ipykernel          6.25.1
ipython            8.14.0
jedi               0.19.0
jupyter_client     8.3.0
jupyter_core       5.3.1
matplotlib-inline  0.1.6
nest-asyncio       1.5.7
outcome            1.2.0
packaging          23.1
parso              0.8.3
passlib            1.7.4
peewee             3.16.2
pexpect            4.8.0
pickleshare        0.7.5
pip                23.2.1
platformdirs       3.10.0
pluggy             1.2.0
prompt-toolkit     3.0.39
psutil             5.9.5
ptyprocess         0.7.0
pure-eval          0.2.2
pyasn1             0.5.0
pycparser          2.21
pydantic           2.1.1
pydantic_core      2.4.0
Pygments           2.16.1
pytest             7.4.0
python-dateutil    2.8.2
python-dotenv      1.0.0
python-jose        3.3.0
python-multipart   0.0.6
pyzmq              25.1.1
requests           2.31.0
rsa                4.9
setuptools         59.6.0
six                1.16.0
sniffio            1.3.0
sortedcontainers   2.4.0
SQLAlchemy         2.0.20
stack-data         0.6.2
starlette          0.27.0
tomli              2.0.1
tornado            6.3.3
traitlets          5.9.0
trio               0.22.2
typing_extensions  4.7.1
urllib3            2.0.4
uvicorn            0.23.2
wcwidth            0.2.6
